+++ src/analyze/src/api/design.js
@@ -2,8 +2,8 @@ import axios from 'axios';
 import utils from '@/utils/utils';
 
 export default {
-  index(model, design) {
-    return axios.get(utils.apiUrl('repos/designs', `${model}/${design}`));
+  index(slug, model, design) {
+    return axios.get(utils.apiUrl('repos/projects', `${slug}/designs/${model}/${design}`));
   },
 
   getTable(table) {