+++ src/analyze/test/unit/specs/views/Repo.spec.js
@@ -1,14 +1,17 @@
 import { mount, createLocalVue, shallowMount } from '@vue/test-utils';
 import Vuex from 'vuex';
+import VueRouter from 'vue-router';
 import Repo from '@/views/Repo';
 import repos from '@/store/modules/repos';
 
 const localVue = createLocalVue();
 localVue.use(Vuex);
+localVue.use(VueRouter);
 
 describe('Repo.vue', () => {
   let actions;
   let state;
+  let router;
   let store;
 
   beforeEach(() => {
@@ -18,6 +21,7 @@ describe('Repo.vue', () => {
       getFile: jest.fn(),
       sync: jest.fn(),
     };
+    router = new VueRouter();
     store = new Vuex.Store({
       modules: {
         repos: {
@@ -31,7 +35,7 @@ describe('Repo.vue', () => {
   });
 
   it('calls getRepo() and sync() via created() lifecycle hook', () => {
-    const wrapper = shallowMount(Repo, { store, localVue });
+    const wrapper = shallowMount(Repo, { store, localVue, router });
 
     expect(wrapper.html()).toBeTruthy();
     expect(actions.getRepo).toHaveBeenCalled();
@@ -39,21 +43,21 @@ describe('Repo.vue', () => {
   });
 
   it('renders no code or markdown by default', () => {
-    const wrapper = mount(Repo, { store, localVue });
+    const wrapper = mount(Repo, { store, localVue, router });
 
     expect(wrapper.element).toMatchSnapshot();
   });
 
   it('renders markdown in the preview pane for markdown files', () => {
     state.activeView = { is_markdown: true, file: '<h1>Title</h1>', populated: true };
-    const wrapper = mount(Repo, { store, localVue });
+    const wrapper = mount(Repo, { store, localVue, router });
 
     expect(wrapper.element).toMatchSnapshot();
   });
 
   it('renders code in the preview pane for code files', () => {
     state.activeView = { is_markdown: false, file: '{ "title": "Title" }', populated: true };
-    const wrapper = mount(Repo, { store, localVue });
+    const wrapper = mount(Repo, { store, localVue, router });
 
     expect(wrapper.element).toMatchSnapshot();
   });
@@ -74,7 +78,7 @@ describe('Repo.vue', () => {
         }],
       },
     };
-    const wrapper = mount(Repo, { store, localVue });
+    const wrapper = mount(Repo, { store, localVue, router });
 
     expect(wrapper.element).toMatchSnapshot();
 
@@ -96,7 +100,7 @@ describe('Repo.vue', () => {
         }],
       },
     };
-    const wrapper = mount(Repo, { store, localVue });
+    const wrapper = mount(Repo, { store, localVue, router });
 
     expect(wrapper.element).toMatchSnapshot();
 
@@ -123,7 +127,7 @@ describe('Repo.vue', () => {
         }],
       },
     };
-    const wrapper = mount(Repo, { store, localVue });
+    const wrapper = mount(Repo, { store, localVue, router });
 
     expect(wrapper.element).toMatchSnapshot();
 
@@ -145,7 +149,7 @@ describe('Repo.vue', () => {
         }],
       },
     };
-    const wrapper = mount(Repo, { store, localVue });
+    const wrapper = mount(Repo, { store, localVue, router });
 
     expect(wrapper.element).toMatchSnapshot();
 
@@ -167,7 +171,7 @@ describe('Repo.vue', () => {
         }],
       },
     };
-    const wrapper = mount(Repo, { store, localVue });
+    const wrapper = mount(Repo, { store, localVue, router });
 
     expect(wrapper.element).toMatchSnapshot();