+++ src/meltano/core/m5o/m5o_file_parser.py
@@ -91,7 +91,6 @@ class MeltanoAnalysisFileParser:
             "many_to_one",
             "many_to_many",
         ]
-        self.project = Project()
 
     def uses_accepted_choice(self, choices, choice):
         return choice in choices